third_party/blink/web_tests/fast/css/zoom-media-queries-width.html

<!DOCTYPE html>
<head>
    <script src="../../resources/js-test.js"></script>
    <style>
        #bla {width: 100px; height: 100px; background: red;}
        @media (max-width: 1200px) {
            #bla {width: 10px; height: 10px; background: blue;}
        }
    </style>
</head>
<body>
<div id=bla></div>

<script>
window.addEventListener("load", function(){
    if (!window.testRunner) {
        document.write("Zoom in manually.");
        return;
    }
    for (var i = 0; i < 5; ++i)
        testRunner.zoomPageIn();

    shouldBeCloseTo('document.getElementById("bla").offsetWidth', 10, 1/64);
}, true);
</script>
</body>
</html>
